+++ exporters/zipkin/zipkin.go
@@ -18,7 +18,6 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -30,12 +29,15 @@ import (
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 )
 
+const (
+	defaultCollectorURL = "http://localhost:9411/api/v2/spans"
+)
+
 // Exporter exports spans to the zipkin collector.
 type Exporter struct {
 	url    string
 	client *http.Client
 	logger *log.Logger
-	config config
 
 	stoppedMu sync.RWMutex
 	stopped   bool
@@ -79,7 +81,8 @@ func WithClient(client *http.Client) Option {
 // New creates a new Zipkin exporter.
 func New(collectorURL string, opts ...Option) (*Exporter, error) {
 	if collectorURL == "" {
-		return nil, errors.New("collector URL cannot be empty")
+		// Use endpoint from env var or default collector URL.
+		collectorURL = envOr(envEndpoint, defaultCollectorURL)
 	}
 	u, err := url.Parse(collectorURL)
 	if err != nil {
@@ -93,6 +96,7 @@ func New(collectorURL string, opts ...Option) (*Exporter, error) {
 	for _, opt := range opts {
 		opt.apply(&cfg)
 	}
+
 	if cfg.client == nil {
 		cfg.client = http.DefaultClient
 	}
@@ -100,7 +104,6 @@ func New(collectorURL string, opts ...Option) (*Exporter, error) {
 		url:    collectorURL,
 		client: cfg.client,
 		logger: cfg.logger,
-		config: cfg,
 	}, nil
 }